+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [2.0.0]
+
+* Change I18n to EasyI18n
+* Change I18nDelegate to EasyI18nDelegate
+* Now EasyI18nDelegate has `supportedLocale`, `localeResolutionCallback` and `localizationsDelegates`
+
 ## [1.0.0]
 
 * I18n